Move files into services directory
