--- __tests__/src/database/queries.js
@@ -1,31 +1,0 @@
-/* eslint-disable */
-
-import queries from '../../../src/database/queries';
-
-describe('queries', () => {
-    'use strict';
-
-    test('findProfiles', () => {
-        // Profiles mock data
-        const profiles = [
-            {
-                // Profile one
-            },
-            {
-                // Profile two
-            },
-        ];
-
-        const realmMock = {
-            objects: jest.fn(),
-        };
-
-        realmMock
-            .objects
-            .mockReturnValueOnce(profiles);
-
-        expect(queries.findProfiles(realmMock)).toBe(profiles);
-
-        expect(realmMock.objects).toBeCalledWith('Profile');
-    });
-});
