+++ src/services/database/schema/v1.js
@@ -1,7 +1,10 @@
 // @flow
 // Schema v1: Effective from 0.3.2 to Current
 
-import {TX_JOB_STATUS_PENDING, TX_JOB_TYPE_NATION_CREATE} from '../../../src/queues/transaction';
+// import {TX_JOB_STATUS_PENDING, TX_JOB_TYPE_NATION_CREATE} from '../../../src/queues/transaction';    // Todo
+
+const TX_JOB_STATUS_PENDING = 200;
+const TX_JOB_TYPE_NATION_CREATE = 'NATION_CREATE';
 
 /**
  * @typedef ProfileType
